Fixed cropping image for person.
git-svn-id: https://svn.java.net/svn/docsources~svn/trunk@2623 3b25252b-017b-4ae5-8f61-fa87a855a72e

+++ DocSources/src/main/java/org/medici/bia/service/peoplebase/PeopleBaseService.java
@@ -769,5 +769,5 @@ public interface PeopleBaseService {
 	 * @param w
 	 * @param h
 	 */
-	public void cropPortraitPerson(Integer personId, Integer x, Integer y, Integer x2, Integer y2, Integer w, Integer h) throws ApplicationThrowable;
+	public void cropPortraitPerson(Integer personId, Double x, Double y, Double x2, Double y2, Double w, Double h) throws ApplicationThrowable;
 }
